Added the minor working changes for the query
